+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * *BREAKING* Update for Split Search & Browse APIs. Refs UIREC-458.
 * Populate internal and external note fields when user sets new piece to "Claim sent". Fixes UIREC-450.
 * Batch rendering of pieces in the receive form. Refs UIREC-461.
+* ECS - Fetch locations, holdings, and items from all related tenants. Refs UIREC-467.
 
 ## [7.0.2](https://github.com/folio-org/ui-receiving/tree/v7.0.2) (2025-04-14)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v7.0.1...v7.0.2)